fix: address review comments
